redesign package:   com.microsoft.azure.toolkit.$platform.$feature[.$subfeature].component/action...   com.microsoft.azure.toolkit.lib.$feature[.$subfeature].service/model/entity...
