--- PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/appservice/intellij/view/AzureForm.java
@@ -1,9 +1,0 @@
-package com.microsoft.azure.toolkit.appservice.intellij.view;
-
-public interface AzureForm<T> {
-    T getData();
-
-    default Object get(String name){
-        throw new RuntimeException();
-    }
-}
