--- PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/appservice/intellij/view/AzureFormInput.java
@@ -1,5 +1,0 @@
-package com.microsoft.azure.toolkit.appservice.intellij.view;
-
-public interface AzureFormInput<T> {
-    T getValue();
-}
