--- PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/appservice/intellij/view/AzureFormPanel.java
@@ -1,7 +1,0 @@
-package com.microsoft.azure.toolkit.appservice.intellij.view;
-
-public interface AzureFormPanel<T> extends AzureForm<T> {
-    void setVisible(boolean visible);
-    default void $$$setupUI$$$() {
-    }
-}
