--- PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/appservice/intellij/view/webapp/WebAppCreationDialog.form
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.azure.toolkit.appservice.intellij.view.webapp.WebAppCreationDialog">
-  <grid id="27dc6" binding="panel" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="316" height="591"/>
-    </constraints>
-    <properties>
-      <maximumSize width="562" height="207"/>
-    </properties>
-    <border type="none"/>
-    <children>
-      <tabbedpane id="d2e82" class="com.intellij.ui.components.JBTabbedPane">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <alignmentX value="0.0"/>
-          <alignmentY value="0.0"/>
-          <doubleBuffered value="true"/>
-          <tabPlacement value="1"/>
-        </properties>
-        <border type="none"/>
-        <children>
-          <grid id="385ea" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-            <margin top="0" left="0" bottom="0" right="0"/>
-            <constraints>
-              <tabbedpane title="Basic"/>
-            </constraints>
-            <properties/>
-            <border type="none"/>
-            <children>
-              <nested-form id="793ad" form-file="com/microsoft/azure/toolkit/appservice/intellij/view/webapp/WebAppConfigFormPanelAdvanced.form"
-                           binding="advancedForm">
-                <constraints>
-                  <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="1" fill="1" indent="0" use-parent-layout="false">
-                    <preferred-size width="295" height="366"/>
-                  </grid>
-                </constraints>
-              </nested-form>
-              <nested-form id="797d" form-file="com/microsoft/azure/toolkit/appservice/intellij/view/webapp/WebAppConfigFormPanelBasic.form" binding="basicForm">
-                <constraints>
-                  <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="1" fill="1" indent="0" use-parent-layout="false">
-                    <preferred-size width="295" height="150"/>
-                  </grid>
-                </constraints>
-              </nested-form>
-            </children>
-          </grid>
-        </children>
-      </tabbedpane>
-    </children>
-  </grid>
-</form>
