--- PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/appservice/webapp/WebAppConfig.java
@@ -1,8 +1,0 @@
-package com.microsoft.azure.toolkit.appservice.webapp;
-
-import com.microsoft.azure.toolkit.appservice.AppServiceConfig;
-import lombok.experimental.SuperBuilder;
-
-@SuperBuilder
-public class WebAppConfig extends AppServiceConfig {
-}
