+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/AzureForm.java
@@ -1,0 +1,9 @@
+package com.microsoft.azure.toolkit.intellij;
+
+public interface AzureForm<T> {
+    T getData();
+
+    default Object get(String name){
+        throw new RuntimeException();
+    }
+}
