+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/AzureFormInput.java
@@ -1,0 +1,5 @@
+package com.microsoft.azure.toolkit.intellij;
+
+public interface AzureFormInput<T> {
+    T getValue();
+}
