+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/AzureFormPanel.java
@@ -1,0 +1,7 @@
+package com.microsoft.azure.toolkit.intellij;
+
+public interface AzureFormPanel<T> extends AzureForm<T> {
+    void setVisible(boolean visible);
+    default void $$$setupUI$$$() {
+    }
+}
