+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/appservice/component/AppServiceConfigFormPanelBasic.form
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.azure.toolkit.intellij.appservice.component.AppServiceConfigFormPanelBasic">
+  <grid id="27dc6" binding="contentPanel" layout-manager="GridLayoutManager" row-count="5" column-count="4" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+    <margin top="0" left="0" bottom="0" right="0"/>
+    <constraints>
+      <xy x="20" y="20" width="417" height="170"/>
+    </constraints>
+    <properties/>
+    <border type="none"/>
+    <children>
+      <component id="b45b0" class="com.intellij.ui.TitledSeparator">
+        <constraints>
+          <grid row="0" column="0" row-span="1" col-span="4" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="true"/>
+        </constraints>
+        <properties>
+          <focusable value="false"/>
+          <font/>
+          <text value="Instance Details"/>
+          <titleFont style="1"/>
+        </properties>
+      </component>
+      <component id="c0640" class="javax.swing.JLabel">
+        <constraints>
+          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="8" fill="0" indent="1" use-parent-layout="true"/>
+        </constraints>
+        <properties>
+          <text value="Name"/>
+        </properties>
+      </component>
+      <component id="d4130" class="javax.swing.JTextField" binding="textName">
+        <constraints>
+          <grid row="1" column="1" row-span="1" col-span="3" vsize-policy="3" hsize-policy="3" anchor="8" fill="1" indent="0" use-parent-layout="true"/>
+        </constraints>
+        <properties/>
+      </component>
+      <component id="c6b66" class="javax.swing.JLabel">
+        <constraints>
+          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="8" fill="0" indent="1" use-parent-layout="true"/>
+        </constraints>
+        <properties>
+          <text value="Runtime stack"/>
+        </properties>
+      </component>
+      <component id="a4196" class="com.microsoft.azure.toolkit.intellij.appservice.component.input.ComboBoxPlatform" binding="selectorPlatform">
+        <constraints>
+          <grid row="2" column="1" row-span="1" col-span="3" vsize-policy="3" hsize-policy="3" anchor="8" fill="1" indent="0" use-parent-layout="true"/>
+        </constraints>
+        <properties/>
+      </component>
+      <component id="3e67" class="com.intellij.ui.TitledSeparator">
+        <constraints>
+          <grid row="3" column="0" row-span="1" col-span="4" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="true"/>
+        </constraints>
+        <properties>
+          <font/>
+          <text value="Deployment"/>
+          <titleFont style="1"/>
+        </properties>
+      </component>
+      <component id="ed0ca" class="javax.swing.JLabel">
+        <constraints>
+          <grid row="4" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="8" fill="0" indent="1" use-parent-layout="true"/>
+        </constraints>
+        <properties>
+          <text value="Application"/>
+        </properties>
+      </component>
+      <component id="7abb2" class="com.microsoft.azure.toolkit.intellij.appservice.component.input.ComboBoxDeployment" binding="selectorApplication">
+        <constraints>
+          <grid row="4" column="1" row-span="1" col-span="3" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties/>
+      </component>
+    </children>
+  </grid>
+</form>
