+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/lib/appservice/webapp/WebAppConfig.java
@@ -1,0 +1,8 @@
+package com.microsoft.azure.toolkit.lib.appservice.webapp;
+
+import com.microsoft.azure.toolkit.lib.appservice.AppServiceConfig;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+public class WebAppConfig extends AppServiceConfig {
+}
